refactor: return deleted uploads from findUploadsCreatedAfter
