+++ packages/db/fauna/resources/Index/uploadIncDeletedSortByCreatedAsc.js
@@ -1,0 +1,31 @@
+import fauna from 'faunadb'
+
+const {
+  CreateIndex,
+  Collection,
+  If,
+  Index,
+  Exists,
+  Not
+} = fauna
+
+/**
+ * Usage:
+ *
+ * Match(Index('upload_inc_deleted_sort_by_created_asc'))
+ */
+const index = {
+  name: 'upload_inc_deleted_sort_by_created_asc',
+  source: Collection('Upload'),
+  values: [
+    { field: ['data', 'created'] },
+    { field: 'ref' }
+  ]
+}
+
+// indexes cannot be updated
+export default If(
+  Not(Exists(Index(index.name))),
+  CreateIndex(index),
+  Index(index.name)
+)
